Revert "Remove duplicate src/dst_stride arguments."
This reverts commit 6c93d8548ca7923510016a70995e527745f18cfd. This
needs reversion because it makes it impossible to support images
where blocksize >= stride, e.g. a 64x64 image using a single 64x64
block. This doesn't fully add support for that case, since the vdsp
function emulated_edge_mc() takes a single (instead of 2) stride
arguments, but at least it doesn't make it impossible to support that
at some point in the future.